Added some integer validation on Add and Modify forms, also fixed a few bugs

+++ D968-InvMngmnt/InHouse.cs
@@ -3,7 +3,7 @@
     public class InHouse : Part
     {
         // Constructor
-        public InHouse(string name, decimal price, int stock, int min, int max, int machine) : base(name, price, stock, min, max, machine)
+        public InHouse(string name, double price, int stock, int min, int max, int machine) : base(name, price, stock, min, max, machine)
         {
             Name = name;
             Price = price;
